refactor: ImageCompressor `util` 패키지로 이동
