+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/attendance/detail/AttendanceDetailViewModel.kt
@@ -15,8 +15,8 @@ import com.yagubogu.ui.attendance.detail.model.AttendanceDetailUiEvent
 import com.yagubogu.ui.attendance.detail.model.CheckInImageItem
 import com.yagubogu.ui.attendance.detail.model.DiaryImageItem
 import com.yagubogu.ui.attendance.detail.model.DiaryMode
-import com.yagubogu.ui.common.component.image.ImageCompressionSpec
-import com.yagubogu.ui.common.component.image.compressImage
+import com.yagubogu.ui.util.ImageCompressionSpec
+import com.yagubogu.ui.util.compressImage
 import com.yagubogu.ui.common.model.PresignedUrlItem
 import com.yagubogu.ui.mapper.toUiModel
 import kotlinx.collections.immutable.ImmutableList
